Divya Drishti/backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
numpy
exifread
piexif
ffmpeg-python
python-magic
httpx
pytest
